requirements.txt
streamlit==1.28.0
streamlit-authenticator==0.2.3
pyyaml==6.0.2
pandas==2.1.3
pymongo==4.6.0
python-dotenv==1.0.1
bcrypt==4.2.1
extra-streamlit-components==0.1.71 